Finalize shape conversion + tests

+++ code/arlab_knowledge/arlab_knowledge/db/entities/entity.py
@@ -2,7 +2,7 @@ from typing import Any, Dict, Optional
 
 import rclpy.logging
 from arlab_knowledge_interfaces import msg
-from sqlalchemy import Float, String
+from sqlalchemy import Float, Integer, String
 from sqlalchemy.orm import (
     Mapped,
     composite,
@@ -39,8 +39,8 @@ class Entity(Base):
 
     stamp: Mapped[TimeData] = composite(
         TimeData._generate,
-        mapped_column("stamp_nanosec", Float),
-        mapped_column("stamp_sec", Float),
+        mapped_column("stamp_nanosec", Integer),
+        mapped_column("stamp_sec", Integer),
     )
 
     shape: Mapped[Optional["Shape"]] = relationship(  # type: ignore # noqa: F821
